+++ js/const.js
@@ -1,0 +1,6 @@
+// Основные константы в проекте
+export const MAX_COMMENTARY_LENGTH = 140;
+export const NUMBER_OF_AVATARS = 6;
+export const NUMBER_OF_PHOTOS = 25;
+export const LIKES_MIN = 15;
+export const LIKES_MAX = 200;
