+++ js/mocks.js
@@ -1,0 +1,34 @@
+// Рандомные данные для генерирования рандомных объектов
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+export const NAMES = [
+  'Леголас',
+  'Фродо',
+  'Гэндальф',
+  'Теодред',
+  'Эомунд',
+  'Сэм',
+  'Арагорн',
+  'Арвен',
+  'Галадриэль',
+  'Эовин',
+];
+export const DESCRIPTIONS = [
+  'Смотрите, куда я приехал!',
+  'Здесь круто!',
+  'Какая вкуснота',
+  'Побеждаю внедорожье!',
+  'Кошки-лепёшки',
+  'Где здесь туалет?',
+  'Дай краба',
+  'Врум-врум',
+  '😐😐😐',
+  '🤷‍♀️',
+  '❤️❤️❤️❤️❤️❤️❤️❤️❤️❤️',
+];
